Moved Vector to own folder
