+++ src/dataStructures/Projectile.ts
@@ -1,6 +1,6 @@
 import { Settings } from "../settings.js"
 import GenericFactory from "./genericObject.js"
-import Vector from "./Vector.js"
+import Vector from "./vector/Vector.js"
 
 export default function Projectile(location: TVector, rotation: Degrees): IRotatableGeneric & ITimeToLive {
