+++ src/dataStructures/vector/Vector.ts
