+++ src/index.ts
@@ -5,7 +5,7 @@ import { Particle } from "./dataStructures/Particle.js"
 import Player from "./dataStructures/Player.js"
 import Position from "./dataStructures/position/Position.js"
 import Projectile from "./dataStructures/Projectile.js"
-import Vector from "./dataStructures/Vector.js"
+import Vector from "./dataStructures/vector/Vector.js"
 import { circleRenderer, playerRenderer, projectileRenderer } from "./draw/composedRenderingFunctions.js"
 import AsteroidSpawnSystem from "./engine/asteroidSpawner.js"
 import global from "./engine/global.js"
